Add/update helper scripts
